predictor_pro/requirements.txt
aiogram==3.10.0
pandas==2.2.0
scipy==1.12.0
python-dotenv==1.0.1
playwright==1.41.0
playwright-stealth==1.0.6
joblib==1.3.2
